src/requirements.txt
requests==2.32.4 ; python_version >= "3.13" and python_version < "4.0"
responses==0.25.8; python_version >= "3.13" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.13" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.13" and python_version < "4.0"
flask-openapi3==4.2.1 ; python_version >= "3.13" and python_version < "4.0"
flask-openapi3-swagger==5.27.1 ; python_version >= "3.13" and python_version < "4.0"
flask-limiter==3.12 ; python_version >= "3.13" and python_version < "4.0"
pytest==8.4.1 ; python_version >= "3.13" and python_version < "4.0"
pytest-cov==6.2.1 ; python_version >= "3.13" and python_version < "4.0"
typeguard==4.4.4 ; python_version >= "3.13" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.13" and python_version < "4.0"
# FIXME: use pysonar when SonarQube instance gets upgraded v10.6+
# (pysonar package requires SonarQube server v10.6 or above)
#pysonar==1.0.2.1722; python_version >= "3.13" and python_version < "4.0"